Refactor climate_baseline_helpers tests
