+++ tests/climate_baseline_helpers/common.py
@@ -1,0 +1,1 @@
+climatology_var_names = ['Tx_Climatology', 'Tn_Climatology', 'Precip_Climatology']
