+++ tests/climate_baseline_helpers/test_climate_baseline_helpers.py
@@ -4,6 +4,8 @@ import struct
 
 from pytest import fixture, mark, raises, fail
 
+from .common import climatology_var_names
+
 from pycds.util import generic_sesh
 from pycds import Network, Station, History, Variable, DerivedValue
 from pycds.climate_baseline_helpers import \
@@ -17,17 +19,8 @@ from pycds.climate_baseline_helpers import \
 
 
 @fixture
-def other_network():
-    return Network(name='Other Network')
-
-@fixture
-def other_climatology_variables(other_network):
-    return [Variable(name=name, network=other_network) for name in climatology_var_names]
-
-
-@fixture
-def sesh_with_other_network_and_climatology_variables(session, other_network, other_climatology_variables):
-    for sesh in generic_sesh(session, [other_network] + other_climatology_variables):
+def sesh_with_other_network_and_climatology_variables(tfs_sesh, other_network, other_climatology_variables):
+    for sesh in generic_sesh(tfs_sesh, [other_network] + other_climatology_variables):
         yield sesh
 
 
@@ -38,44 +31,26 @@ def sesh_with_climate_baseline_variables(sesh_with_other_network_and_climatology
 
 
 @fixture
-def stations():
-    return [Station(native_id=native_id) for native_id in '100 200'.split()]
-
-
-@fixture
-def histories(stations):
-    return [History(
-        station=station,
-        station_name='Station {0}'.format(station.native_id),
-        sdate=datetime.datetime(year, 1, 1),
-        edate=datetime.datetime(year+1, 1, 1),
-    ) for station in stations for year in [2000, 2001]]
-
-
-@fixture
 def sesh_with_station_and_history_records(sesh_with_climate_baseline_variables, stations, histories):
     for sesh in generic_sesh(sesh_with_climate_baseline_variables, stations + histories):
         yield sesh
 
 
-climatology_var_names = ['Tx_Climatology', 'Tn_Climatology', 'Precip_Climatology']
-
-
 def describe_get__or__create__pcic__climate__variables__network():
 
-    def it_creates_the_expected_new_network_record(session):
-        network = get_or_create_pcic_climate_variables_network(session)
-        results = session.query(Network).filter(Network.name == pcic_climate_variable_network_name)
+    def it_creates_the_expected_new_network_record(tfs_sesh):
+        network = get_or_create_pcic_climate_variables_network(tfs_sesh)
+        results = tfs_sesh.query(Network).filter(Network.name == pcic_climate_variable_network_name)
         assert results.count() == 1
         result = results.first()
         assert network.id == result.id
         assert result.publish is True
         assert 'PCIC' in result.long_name
 
-    def it_creates_no_more_than_one_of_them(session):
-        get_or_create_pcic_climate_variables_network(session)
-        get_or_create_pcic_climate_variables_network(session)
-        results = session.query(Network).filter(Network.name == pcic_climate_variable_network_name)
+    def it_creates_no_more_than_one_of_them(tfs_sesh):
+        get_or_create_pcic_climate_variables_network(tfs_sesh)
+        get_or_create_pcic_climate_variables_network(tfs_sesh)
+        results = tfs_sesh.query(Network).filter(Network.name == pcic_climate_variable_network_name)
         assert results.count() == 1
 
 
@@ -126,8 +101,8 @@ def describe_create__pcic__climate__baseline__variables():
         assert result.description == 'Climatological mean of monthly total precipitation'
         assert result.display_name == 'Precipitation Climatology'
 
-    def it_creates_no_more_than_one_of_each(session):
-        sesh = session
+    def it_creates_no_more_than_one_of_each(tfs_sesh):
+        sesh = tfs_sesh
         get_or_create_pcic_climate_baseline_variables(sesh)
         get_or_create_pcic_climate_baseline_variables(sesh)
         results = sesh.query(Variable).filter(Variable.name.like('%_Climatology'))
@@ -272,9 +247,9 @@ def describe_load__pcic__climate__baseline__values():
                     
 def describe_verify__baseline__network__and__variables():
     def describe_without_baseline_network():
-        def it_raises_an_exception(session):
+        def it_raises_an_exception(tfs_sesh):
             with raises(AssertionError) as excinfo:
-                verify_baseline_network_and_variables(session)
+                verify_baseline_network_and_variables(tfs_sesh)
             assert pcic_climate_variable_network_name in str(excinfo.value)
             assert 'not found in database' in str(excinfo.value)
